Added the subjects' files and sign up and login pages
-I have not made the buttons for the sign up and login, but I first
need to figure out why 1) the rectangle or “the banner” is not
centering 2) how to get the buttons (that are not centered) aligned in
a place and stay there.
-There is still a problem with the subject pages going back to the
departments.html page even though now everything should be fine…..it
still shows error in loading page, so we need to figure that out. It
might be fine once we make it into a website with GitHub.